Fix files

+++ init-exercism.sh
@@ -1,6 +1,9 @@
-#! /bin/bash
+#!/bin/bash
 
-sudo echo ''
+if [ `whoami` != 'root' ]; then echo "This program must be run using 'sudo'"; exit; fi
+
+apt update
+apt install -y python3-pip
 
 rm -rf temp
 mkdir temp
@@ -8,9 +11,8 @@ cd temp
 
 wget https://github.com/exercism/cli/releases/download/v3.0.13/exercism-linux-64bit.tgz
 tar -xf exercism-linux-64bit.tgz
-sudo mv exercism /usr/bin
+mv exercism /usr/bin
 exercism version
 
 cd ..
 rm -rf temp
-rm init-exercism.sh
